LayoutTests/imported/w3c/IndexedDB-private-browsing/idbdatabase_deleteObjectStore-expected.txt


PASS IDBDatabase.deleteObjectStore() - object store's name is removed from database's list
